Progress bar seek feature
Current commit contains changes for MOBILE_API and HMI_API.
The progress bar seek feature would allow new
functionality such as the ability to rewind
and fast-forward by selecting the cursor on
the progress bar and dragging it to the desired
place in the song/podcast/audio book. HMI would
only send notifications to the app when desired.
The progress bar would update to show the new track time.