Changed size stack of thread.
Corrected stack size of a thread, which is then used as params for the pthread_attr_setstacksize in Thread::start function.
We are using more than 4K memory for running a test, but set less than 1K.
For the correct work 6K is sufficient, but I set 8K because of precaution.